\name{encounter}
\alias{encounter}
\encoding{UTF-8}
%\alias{overlap.ctmm}
%\alias{overlap.telemetry}
%- Also NEED an '\alias' for EACH other topic documented here.

\title{Calculate the conditional location distribution of ecounters}

\description{To be described in an upcomming manuscript.}

\usage{ encounter(object,include=NULL,exclude=NULL,debias=FALSE,...) }
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{object}{A \code{list} of aligned \code{UD} objects.}
  \item{include}{To be explained.}
  \item{exclude}{To be explained.}
  \item{debias}{Approximate GRF bias correction (in development).}
  \item{...}{Additional arguments for future use.}
}

%\details{}

\value{To be explained.}

%% ~put references to the literature/web site here ~
%\references{}

\author{C. H. Fleming}

%\note{}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{ \code{\link{akde}} }

%\examples{}
