\name{rsf.fit}
\alias{rsf.fit}
\encoding{UTF-8}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Fit integrated resource selection functions (iRSFs) with autocorrelation-adjusted weighted likelihood [IN DEVELOPMENT]}
\description{This function fits integrated resource selection functions with autocorrelation-adjusted weights on the RSF likelihood function, importance sampling, and iterative numerical convergence.
This function is in development and is subject to change.
}
\usage{
rsf.fit(data,UD,R=list(),beta=NULL,integrated=TRUE,isotropic=TRUE,debias=TRUE,smooth=TRUE,
        error=0.01,trace=FALSE,...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{data}{A \code{telemetry} object.}
  \item{UD}{A \code{UD} object generated by \code{\link{akde}} from the same telemetry object as \code{data}. If weights were optimized in \code{\link{akde}}, then they will be adopted by \code{rsf.fit}.}
  \item{R}{A named list of rasters to fit Poisson regression coefficients to (under a log link).}
  \item{beta}{Initial guesstimates for regression coefficients (not currently important).}
  \item{integrated}{Fit an integrated RSF model with simultaneously estimated spatial constraints. (\code{integrated=FALSE} is not yet supported.)}
  \item{isotropic}{Force spatial constraint parameters to be symmetric (Only \code{isotropic=TRUE} is currently supported and this argument will ultimately become depreciated).}
  \item{debias}{Apply a post-hoc bias correction to the spatial constraint parameters.}
  \item{smooth}{Apply location-error smoothing to the tracking data before regression.}
  \item{error}{Numerical error target for the maximized log likelihood. Resulting numerical errors in the parameter estimates will be of a similar relative size. E.g., \eqn{O(1\%)} at \code{error=0.01}.}
  \item{trace}{Report progress on convergence.}
  \item{...}{Arguments passed to \code{optimizer}.}
}

\details{
For autocorrelated tracking data, the relative weights of the log-likelihood used here are taken from the output of \code{\link{akde}}, which are optimzed for non-parametric denstity estimation, and so are approximate here. The absolute weight of the data is taken to be the effective sample size of the integrated spatial parameters, when estimated seperately.

Integrated resource selection functions simultaneously estimate the spatially constraining parameters with the resource selection parameters, rather than first estimating the spatially constraining parameters (usually via MCP) and then holding those parameters fixed---as known values---when estimating the resource selection parameters. The ``integrated'' analysis reduces estimation bias, exposes correlations in the resource and spatial estimate uncertainties, and propagates the spatial estimate uncertainties into the final outputs.

Instead of specifying a number of ``available'' points to sample and having an unknown amount of numerical error to contend with, \code{rsf.fit} specifies an estimation target \code{error} and the number of ``available'' points is increased until this target is met. Moreover, the output log-likelihood is that of the continuous Poisson point process, which does not depend on the number of ``available'' points that were sampled, though the numerical variance estimate is recorded in the \code{VAR.loglike} slot of the fit object.
}

%\value{}

%\section{Warnings}{}

%\references{}

\author{ C. H. Fleming }

%\note{}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{ \code{\link{ctmm.fit}}, \code{\link{optimizer}}, \code{\link{summary.ctmm}}. }

%\examples{}

% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
%\keyword{ ~kwd1 }
%\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
