\name{ctmm::zoom}
\alias{zoom}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Plot with zoom slider}

\description{A generic function to render a plot accompanied with a zoom slider.}

\usage{
zoom(x,...)
}

%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{Object argument.}
  \item{...}{Optional arguments.}
}

%\details{}

%\value{\code{zoom} generates a plot with a zoom slider.}

%\references{}

\author{C. H. Fleming}

%\note{}

%% ~Make other sections like Warning with \section{Warning }{....} ~

%\seealso{ \code{\link{akde}}, \code{\link{as.telemetry}}.  }

%\examples{\donttest{
%# Load package and data
%library(ctmm)
%data(buffalo)

%# Extract movement data for a single animal
%billy <- buffalo[[1]]

%# Plot the data
%plot(billy)
%}}

% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{internal}
%\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
