\name{rsf.fit}
\alias{rsf.fit}
\encoding{UTF-8}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Fit integrated resource selection functions (iRSFs) with autocorrelation-adjusted weighted likelihood [IN DEVELOPMENT]}
\description{This function fits integrated resource selection functions with autocorrelation-adjusted weights on the RSF likelihood function, importance sampling, and iterative numerical convergence.
This function is in development and is subject to change.
}
\usage{
rsf.fit(data,UD,R=list(),mod=NULL,int=TRUE,iRSF=TRUE,level.UD=0.99,isotropic=TRUE,
        beta=NULL,debias=TRUE,smooth=TRUE,error=0.01,trace=TRUE,...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{data}{A \code{telemetry} object.}
  \item{UD}{A \code{UD} object generated by \code{\link{akde}} from the same telemetry object as \code{data}. If weights were optimized in \code{\link{akde}}, then they will be adopted by \code{rsf.fit}.}
  \item{R}{A named list of rasters to fit Poisson regression coefficients to (under a log link).}
  \item{mod}{Effect modifier vector or list to incorporate potentially time-dependent effect modifiers with the rasters in \code{R}, without increasing memory costs. Each entry must be an effect modifier column name in \code{data}, or a list of such column names. (Not supported yet.)}
  \item{int}{Intercept vector to disable intercept terms, if \code{int=FALSE}, in the case of effect modifiers provided in argument \code{mod}.}
  \item{iRSF}{Fit an integrated RSF model with simultaneously estimated spatial constraints. \code{iRSF=FALSE} is for comparison purposes only.}
  \item{level.UD}{Coverage probability of \code{UD} to sample uniformly from if \code{iRSF=FALSE}. Can also be a pre-defined spatial polygon object.}
  \item{isotropic}{Force spatial constraint parameters to be symmetric (only \code{isotropic=TRUE} is currently supported).}
  \item{beta}{Initial guesstimates for regression coefficients (not currently important).}
  \item{debias}{Apply a post-hoc bias correction to the spatial constraint parameters.}
  \item{smooth}{Apply location-error smoothing to the tracking data before regression.}
  \item{error}{Relative numerical error threshold for the parameter estimates and log-likelihood.}
  \item{trace}{Report progress on convergence (see Details).}
  \item{...}{Arguments passed to \code{optimizer}.}
}

\details{
For autocorrelated tracking data, the relative weights of the log-likelihood used here are taken from the output of \code{\link{akde}}, which are optimzed for non-parametric denstity estimation, and so are approximate here. The absolute weight of the data is taken to be the effective sample size of the integrated spatial parameters, when estimated seperately.

Integrated resource selection functions simultaneously estimate the spatially constraining parameters with the resource selection parameters, rather than first estimating the spatially constraining parameters (usually via MCP) and then holding those parameters fixed---as known values---when estimating the resource selection parameters. The ``integrated'' analysis reduces estimation bias, exposes correlations in the resource and spatial estimate uncertainties, and propagates the spatial estimate uncertainties into the final outputs.

Instead of specifying a number of ``available'' points to sample and having an unknown amount of numerical error to contend with, \code{rsf.fit} specifies an estimation target \code{error} and the number of ``available'' points is increased until this target is met. Moreover, the output log-likelihood is that of the continuous Poisson point process, which does not depend on the number of ``available'' points that were sampled, though the numerical variance estimate is recorded in the \code{VAR.loglike} slot of the fit object.

When \code{trace=TRUE}, a number of convergence estimates are reported, including the standard deviation of the numerical error of the log-likelihood, SD[\eqn{\log(\ell)}], the most recent log-likelihood update, d\eqn{\log(\ell)}, and the most recent (relative) parameter estimate updates d\eqn{\hat{\beta}/}SD[\eqn{\hat{\beta}}].
}

%\value{}

%\section{Warnings}{}

%\references{}

\author{ C. H. Fleming }

%\note{}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{ \code{\link{ctmm.fit}}, \code{\link{optimizer}}, \code{\link{summary.ctmm}}. }

%\examples{}

% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
%\keyword{ ~kwd1 }
%\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
